requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
pydantic-settings
python-dotenv
PyJWT
python-multipart
requests
openai
APScheduler
pandas
numpy
yfinance